requirements.txt
aiohttp>=2.1.0
async-timeout>=1.2.1
CacheControl>=0.12.3
certifi>=2017.4.17
cffi>=1.10.0
chardet>=3.0.3
click>=6.7
colorama>=0.3.9
colorlog>=2.10.0
discord.py>=0.16.8
idna>=2.5
imdbpie>=4.2.0
lockfile>=0.12.2
msgpack-python>=0.4.8
multidict>=2.1.6
pip>=9.0.1
py>=1.4.34
Pybooru>=4.1.0
pycparser>=2.17
PyNaCl>=1.1.2
pytest>=3.1.1
Python-Trivia-API>=1.0.1
PyYAML>=3.12
requests>=2.17.3
six>=1.10.0
urllib3>=1.21.1
websockets>=3.3
xmltodict>=0.11.0
yarl>=0.10.2
